Webpack building index.js in dist
