+++ package.json
@@ -7,6 +7,9 @@
     "lib": "lib"
   },
   "devDependencies": {
+    "babel-core": "^6.26.3",
+    "babel-loader": "^7.1.2",
+    "babel-preset-env": "^1.7.0",
     "webpack": "^3.10.0"
   },
   "scripts": {
