+++ visualiser/index.html
@@ -4,23 +4,13 @@
 <link rel="stylesheet" type="text/css" href="styles.css">
 
 <!-- WorkFlo-->
-<script src="workflo/node.js"></script>
-<script src="workflo/nodecontainer.js"></script>
-<script src="workflo/connector.js"></script>
-<script src="workflo/workflo.js"></script>
-<link rel="stylesheet" type="text/css" href="workflo/workflo.css">
+<script src="../lib/workflo/node.js"></script>
+<script src="../lib/workflo/nodecontainer.js"></script>
+<script src="../lib/workflo/connector.js"></script>
+<script src="../lib/workflo/workflo.js"></script>
+<link rel="stylesheet" type="text/css" href="../lib/workflo/workflo.css">
 
-<!-- BT Test -->
-<script src="js/nodes/root.js"></script>
-<script src="js/nodes/flip.js"></script>
-<script src="js/nodes/action.js"></script>
-<script src="js/nodes/condition.js"></script>
-<script src="js/nodes/selector.js"></script>
-<script src="js/nodes/sequence.js"></script>
-<script src="js/nodes/lotto.js"></script>
-<script src="js/nodes/repeat.js"></script>
-<script src="js/nodes/wait.js"></script>
-<script src="js/mistreevous.js"></script>
+<script src="../dist/index.js"></script>
 
 <head>
 	<meta charset="utf-8">
